Moved Path classes to position.py
