+++ utilities/position.py
@@ -1,6 +1,6 @@
 from math import pi
 
-# from dataclasses import dataclass
+from dataclasses import dataclass
 from wpimath.geometry import Rotation2d, Translation2d, Pose2d
 
 # Is this redundant? -> Using Pose2d instead
@@ -8,6 +8,45 @@ from wpimath.geometry import Rotation2d, Translation2d, Pose2d
 # class NodePosition:
 #     translation: Translation2d
 #     heading: Rotation2d
+
+
+@dataclass
+class Path:
+    waypoints: list[Translation2d]
+    final_heading: Rotation2d
+
+    def copy(self):
+        return Path(
+            [Translation2d(i.x, i.y) for i in self.waypoints],
+            Rotation2d(self.final_heading.radians()),
+        )
+
+    def __add__(self, other):
+        return Path(self.waypoints + other.waypoints, self.final_heading)
+
+    def __iter__(self):
+        return iter(self.waypoints)
+
+    def __getitem__(self, __i: int | slice):
+        return self.waypoints[__i]
+
+    def __setitem__(self, __key: int, __value: Translation2d):
+        self.waypoints[__key] = __value
+
+    def __delitem__(self, __key: int | slice):
+        del self.waypoints[__key]
+
+    def __len__(self):
+        return len(self.waypoints)
+
+
+@dataclass
+class NotePaths:
+    # All paths assume RED alliance
+    # They will automatically be flipped if we are blue
+    pick_up_path: Path
+    shoot_path: Path
+    pickup_offset: Translation2d
 
 
 # StageHeight = 8.210550308227539
